refactor: hickey + lowy over the diff
**Concept multiplication (hickey L3, the survey's hit).** The diff added a
second "a note, as one dim clamped line under a title" abstraction beside
the canonical one — `filter/excerpt.ts` next to `note/preview.ts`, same
slot, same clamp, differing only in which line is taken and whether the
marks are dropped. Under the alternative-placement test the generative
side of "a note reduced to one line under a title" is the note, not the
filter, so the excerpt moves to `note/excerpt.ts` / `note/Excerpt.tsx`
and the two files now say what they share and why the difference is real
rather than a parameter (a preview is about an OPENING and may tidy; an
excerpt is about a POSITION, and dropping characters would move the hit
away from the offsets the matcher found it at). `filter/lit.ts` stays put:
it is the query half both a title and a note draw through.

**Lowy §4, symmetry.** Three surfaces draw a narrowed row and each draws
a different subset. Two of the three absences were argued at the site
already (a day keeps no context rows; a day's rows are records, not
placements); the third was silent — a trash row has no note body, so
there is no ¶ for a note-only hit to be excerpted from. Named where it
would otherwise read as a missed axis.

**Fragmentation (hickey L2): no invariants found.** `lighting`,
`asContext` and `behindTheMark` are three total projections of ONE value
— the narrowing's `matched` map — with no rule holding them together and
no reconciliation machinery between them: a row cannot be context and lit
at once because both answers read the same lookup. Enumerated rather than
assumed.
