+++ docs/brainstorming/filter-in-place.md
@@ -219,7 +219,7 @@ The fix is three things, all view-time — nothing is stored, nothing is re-matc
 
 - **The words are lit where they sit.** `needlesOf` hands a parsed query's positive words to the view and `litBy` says where each lands in a text; both are in `@olai/format`'s `filter.ts`, beside the fold they use, which is the whole point — an offset found under some other case rule would light up a stretch of title the matcher never looked at. The two title paths that are already held byte-for-byte to each other (`markdown/plain.ts`'s fast path and the HAST walk) take the needles through one split (`filter/lit.ts`), so a tag pill lights inside itself rather than instead of itself, and `plain.test.ts` sweeps both readings.
 - **A row that did not match is dim** — the same utility a row that cannot be started yet already wears (`blocked.ts`'s `WAITING_DIM`), deliberately the same value rather than a second number, since a row can be both at once. On the LINE and the body, never on the `<li>`: opacity compounds through a subtree, and an item would dim the very match the context leads to.
-- **A note-only match draws one clamped line** of its note, opening at the hit's own line, with the word lit (`filter/excerpt.ts`). It replaces the density preference's top-of-note preview on that row rather than sitting beside it.
+- **A note-only match draws one clamped line** of its note, opening at the hit's own line, with the word lit (`note/excerpt.ts`, beside `note/preview.ts` — one concept, two selectors). It replaces the density preference's top-of-note preview on that row rather than sitting beside it.
 
 **Auto-expanding a matched note lost, and it was the obvious alternative.** Notes here run to ~1.5KB paragraphs; the filter re-evaluates per keystroke, so a page of open notes would reflow violently under somebody still typing; and it would trample the reader's own open/closed state, which would then need saving and restoring to put back. A clamped window is the same idea with a bounded cost.
