+++ packages/web/README.md
@@ -397,7 +397,7 @@ The fourth door onto the one query language, and the only one that does not list
 - `narrowed.tsx` — the WAY to that reading from wherever a row is drawn: a `Pick` of it rather than a second declaration, carried by a context, with a default of "nothing is filtered" for a `<Tree>` drawn outside the provider.
 - `why.ts` — what a row ASKS of it, which is the whole of "a filtered row says why it is drawn" (2026-08-18): `matchedAttr` (the one spelling of `data-match`, now that a tree row, a day row and a trash row each draw it), `lighting` (the words to highlight in a matched row's title), `CONTEXT_DIM` (a row drawn only as the ancestry leading somewhere, in the same utility `blocked.ts` dims a waiting row with — one value, so a row that is both does not compound), `behindTheMark` (the words to excerpt on a row whose only hit is in its note, which is a `Match.field` of `desc` and nothing this file re-derives), and `unfiltered`, which every "this page is empty" sentence is drawn on. Pure, and beside the context rather than in it, because three surfaces and a test ask it directly.
 - `lit.ts` — one text as the runs the query landed on and the runs it did not. WHERE they land is `@olai/format`'s `litBy`, over the same fold `matching` searched with, so a highlight cannot appear in a stretch of text the matcher never looked at. Runs rather than markup because there are three writers: the two title paths that are held byte-for-byte to each other (`markdown/plain.ts`'s fast path and `markdown/title.ts`'s HAST walk, both through `markdown/tags.ts`), and a note's excerpt, which is Solid elements where the escaping is the framework's.
-- `excerpt.ts` / `Excerpt.tsx` — the one clamped line of a note a row was found BY, opening at the hit's own line, cut from the landings themselves rather than searched for a second time inside the window. Deliberately NOT auto-expanding the note: ~1.5KB paragraphs, a filter that re-evaluates per keystroke, and a reader's own open/closed state to trample. It replaces the density preference's top-of-note preview on that row rather than sitting beside it.
+- **The one clamped line of a note a row was found BY** is not here: it is `note/excerpt.ts` and `note/Excerpt.tsx`, beside `note/preview.ts`, because the two are one concept with two selectors — the same slot under a title and the same clamp, differing only in which line and whether the marks are dropped. What is here is the query half they both draw through (`lit.ts`). Deliberately NOT auto-expanding the note: ~1.5KB paragraphs, a filter that re-evaluates per keystroke, and a reader's own open/closed state to trample.
 - **Folds are suspended while a filter is on**, and that lives in `fold/reading.ts` rather than here: a fold is a claim about the tree the reader was reading, and honouring one inside a filtered tree would hide the match the filter was typed to find. The memory is untouched; clearing the filter restores every collapse. It is one accessor because FOUR things walk it — the tree draws from it, and the editor, the selection and the drag walk it for where `↑`/`↓` go, what a shift-click spans and where a drop can land (`edit/Editable.tsx` hands one to all three). A reading the tree suspended and those three did not is a page whose arrow keys walk rows nobody can see.
 - `FilterBar.tsx` — the box, the count, the clear, and the refusal line. A known operator with an unknown value is REPORTED in the grammar's own words, never quietly searched for as text.
 - `tag.ts` — a `#tag` pressed. A delegated listener on the main pane, because a tag pill arrives through `innerHTML` (`markdown/tags.ts`) and belongs to no component — the same situation, and the same answer, as a link inside rendered markdown (`router.tsx`'s `followed`). The row's own title click declines a press that landed on a pill, or one press would filter the page AND open an editor. A pill inside an ancestry crumb goes where the crumb goes: the walk skips anchors, so the link has already answered.
