+++ packages/web/src/client/NodeBody.tsx
@@ -60,8 +60,8 @@ import { docOf, type LocatedRegular } from "@olai/format"
 import { createMemo, Show } from "solid-js"
 
 import { DocRef } from "./document/DocRef.tsx"
-import { Excerpt } from "./filter/Excerpt.tsx"
-import { excerptOf } from "./filter/excerpt.ts"
+import { Excerpt } from "./note/Excerpt.tsx"
+import { excerptOf } from "./note/excerpt.ts"
 import { plainLine } from "./note/preview.ts"
 import { Note } from "./Note.tsx"
 import { PropsDrawer } from "./props/PropsDrawer.tsx"
